requirements.txt
requests
beautifulsoup4
rapidfuzz
rich
python-dotenv
streamlit
